fix import cycle not allowed problem
